change: scripts/**/*.js -> .mjs
