+++ scripts/generator/copyByJson.mjs
