+++ scripts/manager/modules/postEmojiAlias.mjs
