+++ scripts/manager/modules/pretender.mjs
@@ -1,17 +1,16 @@
 import puppeteer from "puppeteer";
-import { goToEmojiPage } from "./goToEmojiPage";
-import { getLocalJson } from "./getLocalJson";
-import { postEmojiAdd } from "./postEmojiAdd";
-import { outputLogJson } from "../../utilities/outputLogJson";
-import { outputResultJson } from "../../utilities/outputResultJson";
+import { goToEmojiPage } from "./goToEmojiPage.mjs";
+import { getLocalJson } from "./getLocalJson.mjs";
+import { postEmojiAlias } from "./postEmojiAlias.mjs";
+import { outputLogJson } from "../../utilities/outputLogJson.mjs";
+import { outputResultJson } from "../../utilities/outputResultJson.mjs";
 
-export const uploader = async (inputs) => {
+export const pretender = async (inputs) => {
   const {
     browser: BROWSER,
     configs: CONFIGS,
     debug: DEBUG,
     log: LOG,
-    excludeExplicit: EXCLUDE_EXPLICIT,
     term: TERM,
     time: TIME,
   } = inputs;
@@ -19,39 +18,34 @@ export const uploader = async (inputs) => {
   let i = 0; // 再帰でリストの続きから処理するためにインデックスを再帰関数の外に定義する
   let FAILED = false;
   let RELOGIN = false;
-  const rawLocalDecomojiList = getLocalJson(
+  const localDecomojiList = getLocalJson(
     CONFIGS,
     TERM,
-    ["upload"],
-    "uploder",
+    ["rename"],
+    "pretender",
     LOG
   );
-  // バージョンごとに追加するとき、excludeExplicit=true なら explicit デコモジを取り除く
-  const localDecomojiList =
-    TERM === "version" && EXCLUDE_EXPLICIT
-      ? rawLocalDecomojiList.filter(
-          ({ path }) => !RegExp("explicit").test(path)
-        )
-      : rawLocalDecomojiList;
   const localDecomojiListLength = localDecomojiList.length;
 
   TERM === "version" &&
     LOG &&
-    outputLogJson(localDecomojiList, "filtered", "uploder");
+    outputLogJson(localDecomojiList, "list", "pretender");
 
   const result = {
     error: [],
+    error_invalid_alias: [],
     error_name_taken: [],
     error_name_taken_i18n: [],
     ok: [],
   };
   const messages = {
-    ok: "uploaded",
+    ok: "registered",
+    error_invalid_alias: "skipped(target no exists)",
     error_name_taken: "skipped(already exists)",
     error_name_taken_i18n: "skipped(international emoji set already includes)",
   };
 
-  const _upload = async (inputs) => {
+  const _pretend = async (inputs) => {
     // puppeteer でブラウザを起動する
     const browser = await puppeteer.launch({
       devtools: BROWSER,
@@ -72,23 +66,24 @@ export const uploader = async (inputs) => {
       return inputs;
     }
 
-    TIME && console.time("[Upload time]");
+    TIME && console.time("[Register time]");
     while (i < localDecomojiListLength) {
-      const { name, path } = localDecomojiList[i];
-      // name か path が falsy の時は FAILED フラグを立ててループを抜ける
-      if (!name || !path) {
+      const { name, alias_for } = localDecomojiList[i];
+      // name か alias_for が falsy の時は FAILED フラグを立ててループを抜ける
+      if (!name || !alias_for) {
         FAILED = true;
         break;
       }
 
-      const res = await postEmojiAdd(page, WORKSPACE, name, path);
+      const res = await postEmojiAlias(page, WORKSPACE, name, alias_for);
 
       console.info(
         `${i + 1}/${localDecomojiListLength}: ${
           res.ok
             ? messages.ok
             : res.error === "error_name_taken" ||
-              res.error === "error_name_taken_i18n"
+              res.error === "error_name_taken_i18n" ||
+              res.error === "error_invalid_alias"
             ? messages[res.error]
             : res.error
         } ${name}.`
@@ -98,7 +93,8 @@ export const uploader = async (inputs) => {
       res.ok
         ? result.ok.push(name)
         : res.error === "error_name_taken" ||
-          res.error === "error_name_taken_i18n"
+          res.error === "error_name_taken_i18n" ||
+          res.error === "error_invalid_alias"
         ? result[res.error].push(name)
         : res.error === "ratelimited" // ratelimited エラーの場合はログに残さない
         ? void 0
@@ -121,7 +117,8 @@ export const uploader = async (inputs) => {
       if (
         res.error &&
         res.error !== "error_name_taken" && // 登録済みのエラー
-        res.error !== "error_name_taken_i18n" // i18n と競合するエラー
+        res.error !== "error_name_taken_i18n" && // i18n と競合するエラー
+        res.error !== "error_invalid_alias" // エイリアスを貼る先が見つからないエラー
       ) {
         FAILED = true;
         break;
@@ -133,7 +130,7 @@ export const uploader = async (inputs) => {
       FAILED = false;
       RELOGIN = false;
     }
-    TIME && console.timeEnd("[Upload time]");
+    TIME && console.timeEnd("[Register time]");
 
     // ブラウザを閉じる
     if (!DEBUG) {
@@ -144,19 +141,19 @@ export const uploader = async (inputs) => {
     if (RELOGIN) {
       TIME && console.timeLog("[Total time]");
       console.info("Reconnecting...");
-      return await _upload(inputs);
+      return await _pretend(inputs);
     }
 
     // 追加中に ratelimited にならなかった場合ここまで到達する
     if (FAILED) {
-      console.error("[ERROR]Upload failed.");
+      console.error("[ERROR]Register failed.");
     }
-    console.info("Upload completed!");
-    outputResultJson(result, "result", "uploder");
+    console.info("Register completed!");
+    outputResultJson(result, "result", "pretender");
     // 入力し直したかもしれないので返す
     return inputs;
   };
 
   // 再帰処理をスタートする
-  return await _upload(inputs);
+  return await _pretend(inputs);
 };
