+++ scripts/manager/modules/recursiveInputAccount.mjs
@@ -1,6 +1,6 @@
 import inquirer from "inquirer";
-import { isEmail } from "../../utilities/isEmail";
-import { isInputs } from "../../utilities/isInputs";
+import { isEmail } from "../../utilities/isEmail.mjs";
+import { isInputs } from "../../utilities/isInputs.mjs";
 
 // ログインエラーの時の再帰処理
 export const recursiveInputAccount = async (browser, page, inputs) => {
