+++ scripts/utilities/convertBufferToArray.mjs
