+++ scripts/utilities/convertFilepathToBasename.mjs
