+++ scripts/utilities/convertToLowerCasedArray.mjs
@@ -1,4 +1,4 @@
-import { isStringOfNotEmpty } from "./isStringOfNotEmpty";
+import { isStringOfNotEmpty } from "./isStringOfNotEmpty.mjs";
 
 // 配列のアイテムを小文字に変換して返す。文字列でない場合は何もしない
 export const convertToLowerCasedArray = (array) => {
