+++ scripts/utilities/getDecomojiCategory.mjs
