+++ scripts/utilities/getMergedDiffOfCategories.mjs
