+++ scripts/utilities/getMergedDiffOfManages.mjs
