+++ scripts/utilities/isDecomojiFile.mjs
