+++ scripts/utilities/isEmail.mjs
