+++ scripts/utilities/isStringOfNotEmpty.mjs
