+++ scripts/utilities/writeJsonFileSync.mjs
